docs(#3630): scope statements at both unreachable sites; split row construction out of value.rs
Lead ruling on coord-request 3630-R2 batched the two roborev Lows and imposed two
conditions. Condition (1): record the unreachability as a SCOPE STATEMENT in code
at both sites WITH the measurement — never 'TODO: add test', because a reader must
be able to tell unreachable-measured from nobody-got-round-to-it. Both sites now
carry the routes tried, the cause, and the TRIGGER that would make them testable.

Adding them pushed value.rs to 825/800 and the file-size ratchet FAILED. The file
was 660 on main, so my change caused it and waiving with CQLITE_ALLOW_FILE_GROWTH
would have been the wrong call. Split BY RESPONSIBILITY instead (#1116): row
construction — ColumnKeys, intern_column_keys, row_to_object — moves verbatim to
bindings/node/src/row.rs, beside its own mechanism module row_properties.rs, while
value.rs keeps CQL value conversion. value.rs is now 631, BELOW its own pre-change
baseline. ConvCtx::env is now pub(crate); two call sites re-pointed.

Verified after the split: 163 Node tests pass (issue-3630, issue-3504, types,
typescript-definitions), Python 5/5, clippy -D warnings clean, file-size PASS.
